Move diagnostic rendering to CLI, unify dotted module path resolution in ImportTable
