+++ src/cli/emit.rs
@@ -37,13 +37,13 @@ pub fn cmd_emit(file: &str, target: &str, emit_ast: bool, emit_ir: bool, no_chec
             .collect();
         if !errors.is_empty() {
             for d in &errors {
-                eprintln!("{}", d.display_with_source(&source_text));
+                eprintln!("{}", crate::diagnostic_render::display_with_source(d, &source_text));
             }
             eprintln!("\n{} error(s) found", errors.len());
             std::process::exit(1);
         }
         for d in diagnostics.iter().filter(|d| d.level == diagnostic::Level::Warning) {
-            eprintln!("{}", d.display_with_source(&source_text));
+            eprintln!("{}", crate::diagnostic_render::display_with_source(d, &source_text));
         }
         checker_opt = Some(checker);
     }
